network: ovs host with uninitialized data center
Waiting for the data center to reach UP status after moving a host to a
different cluster encapsulates the hidden assumption that there is a
storage domain in the data center. This might not be the case, e.g.
during debug, and is a redundant assumption.
On the other hand, the fixture should wait for host networks to be in
sync after trying to sync them, in order to hand off a complete
operation.
In test_ovn_physnet the vm_down fixture was unaware of relying on the
above assumption in host_in_ovs_cluster fixture. Now, make the
dependency on a storage domain explicit and where it is directly needed.

Change-Id: Id53f1adb674f6d5e1c05f2e88544e9ee691979b0
Signed-off-by: Eitan Raviv <eraviv@redhat.com>