network: ovn - wait for port security to take effect
Enabling port security involves communication with the ovn provider
which then needs to apply the request back to the switch. Pinging a VM
immediately after sending the request creates a race between the request
from the provider to the switch and the ping request. Therefore add a
sleep after sending the request to the provider and before testing its
application with a ping.

Change-Id: I9d8f69ee0f990f6583de5c53674c49d3ee787a33
Signed-off-by: Eitan Raviv <eraviv@redhat.com>