_detect_problematic_hosts: add Connecting state
We can actually get a Connecting state during host installation. Once
the 120s timeout passes we finish installation and move to Initializing
state. If a network exception occures (e.g. when vdsm is not up yet)
when we try to get capabilities we reach the Connecting flow since the
installation has already finished

Change-Id: Ide94c6bb0a3fe9696e459360f87b0099436f66e4
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>