licensing: check copyright notices
Require new patches to include SPDX tags instead of the old
copyright/license notice.

It's enough to include the following two lines, in every file that
should be copyrighted:

Copyright oVirt Authors
SPDX-License-Identifier: GPL-2.0-or-later

The last one should be included (perhaps with a different license, if
needed) also if for any reason a copyright notice is still needed in the
file, and also if the copyright line should be different. E.g. if a file
is copied from some other project, with its own copyright and license.

adapted from https://gerrit.ovirt.org/#/c/ovirt-system-tests/+/112501/

Change-Id: Id98078b140e7ef4b190760457494f2b1bb35bfa8
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>