engine: Change the default engine password
With the default engine password being "123" there's actually
a significant chance of it being a substring of the authentication token
returned by the engine. That, combined with usage of ansible collection,
means it will be replaced with "*******" characters [1] breaking
the authentication process. We actually experienced failures like these
in CI.

Let's replace the default password to the one that we already use throughout
the rest of OST to minimize the chance of it appearing in the token.

[1] https://github.com/ansible/ansible/issues/23671

Change-Id: I2795a7f75574e25955af8b96ce122d7be9aec18f
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>