ansible: Adapt to ipv6-{first,exclusive} runs
This patch adapts the ansible to the ipv6-exclusive and ipv6-first
approach that we currently use in OST.

The nfs storage domains are now using fqdn instead of plain ip
addresses.

For 'ovirt_auth' and the 'infra' role we need to stick with
'engine_ip_url' for now since the ansible executor image cannot reach
the engine VM with fqdn yet. This will require a deeper refactoring and
will be implemented in subsequent patches.

Change-Id: If2b6da1f98278ed8ff4f78dd00264b9afb903428
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>