parametrize ansible execution environment image
We use by default an image that we build nightly from released 
ovirt-ansible-collection - quay.io/ovirt/el8stream-ansible-executor.
For testing ansible patches it's convenient to use a different image on
request, so we introduce OST_ANSIBLE_EEI environment variable that can
be used to pull and use a custom image.

Change-Id: Ic03c75bec4616f3ec865667564aa7831759b5748
Signed-off-by: Martin Necas <mnecas@redhat.com>