engine: Get rid of hanging systemd-coredump
From time to time we find a 'systemd-coredump' process hanging
indefinitely on the engine and eating 100% of one of the CPUs.  It seems
that it tries to collect a coredump from a crashed
'systemd-cgroups-agent' process.  The crash itself happens ~20 secs
after the VM started, so it's most probably not OST nor oVirt-related.

The problem manifests itself much later with tests failing due to slow
storage response time from the engine VM caused by the constant CPU
spike.

Let's investigate the problem on the side and simply kill the
'systemd-coredump' process if we find one for now.

Change-Id: I3599c7c0cdb6187a5e9af83701ed2b6cd01b5095
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>