he: skip updating ovirt-release-master-tested
we wipe out all repo files and use the ones from host-0, but the rpm
database may still contain our release rpms that can get updated and
pull in those repo files again. We generally do not want that as we just
want to "see" the supplied custom repositories. It's not technically
needed for other suites, but it shouldn't hurt so let's skip updating
"tested" rpm everywhere.

Change-Id: I467f7a0e76ede65ba673015bd46cdbb5364b5a93
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>