bootstrap: Allow numa node data to settle
From time to time OST fails the 'test_get_host_numa_nodes' test
complaining that there are no numa nodes on the host.
From an analysis of the 'engine.log' it seems that a host may
go into 'UP' state, but only after that 'RefreshHostDevicesCommand'
is triggered and the numa data, among other, is filled for the host.

Let's use retries and move the test to execute a bit later to allow
the data to settle.

Change-Id: I28d6d342d77afbc7c5144eebb448fc7aee4912b1
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>