engine_setup: enable cinderlib by default
we want to enable it so we test the integration that's now included in
oVirt by default. It is NOT enabled in HE suites since the configuration
of HE is not customized. This way we test both with and without - which
is ok until we create a real cinderlib/ceph test.

Change-Id: I45bae4a7660d6b03645eb15af5d01d8f786a2342
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>