network: test port isolation - use serial vmconsole
The test relied on the VMs to get ips from dnsmasq on the isolated
network - eth1 in the VM. This proved flaky.
This patch uses the serial vmconsole to check if an ip was assigned by
dhcp to eth1 on the VMs, and if not, invokes a dhcp request to get an
ip.

Change-Id: Ic218c5692f11826b79d8332f34b6cf597e789280
Signed-off-by: Eitan Raviv <eraviv@redhat.com>