lagofy: run IPv6 to IPv4 sshd SOCKS proxy
this is needed in IPv6-only networks since resources.ovirt.org and other
things don't work over IPv6, this way we use the OST executor as a SOCKS
proxy that allows VMs to dnf update.

sshd starts in ost_init, listens on port 2222 of management network interface only, and gets shut down in
ost_destroy. On VMs side a sshd_proxy service is added and enabled that starts ssh client opening a SOCKS
proxy on port 1234, and finally DNF is reconfigured to use that proxy.

In case of HE we need to tunnel all the way via host-0, so we create a plain tunnel from HE VM to host-0
and connect it to port 1234 locally.

Change-Id: Ie43c5bb4fe41310b8e389a32a8e194a7379e8607
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>