Kill hanging systemd-coredump processes at boot
From time to time we see problems with systemd-cgroups-agent
crashing with SIGABRT and systemd-coredump hanging indefinitely
eating 100% cpu while trying to generate a coredump from that crash.
This was happening only on the engine/storage VMs until now because
hosts were using abrt for coredump generation. Currently abrt
is not effective on the hosts, so we need to look for these
issues there as well.

Change-Id: Ie7ec74c39f349856c369f12c4889a1376e0398ed
Signed-off-by: Andrej Cernek <acernek@redhat.com>