deployment: Make custom repos working on Node hosts
The Node hosts prevent package upgrades.  This is intended in Node but
it also means that packages provided in --custom-repo ost.sh command
line option won't be used.

Let's remove the Node package version lock when a custom repo is
requested.  And let's still keep ovirt-node-ng-image-update version
because upgrading this package takes a looong time.

This is not going to work with tests running Node upgrades.  Such
tests shouldn't use custom repos.

There are some general arguments for and against this change:

- Node hosts are not really designed to have individual packages
  upgraded.  OST gating and other OST runs with custom repos should
  use non-Node hosts.  Node hosts should be tested only in OST loop
  runs, not on regular CI.

- Packages on Node hosts can still be upgraded, it works and Node
  nightly has the same initial set of packages as non-Node nightly.
  It is safe to use a Node host under such circumstances and it
  reduces the number of environments that must be tested separately.

Change-Id: Icd7dabc8619232183f764e11181d574a360dd410
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>