network: import cirros template with imageio
An IPv6 only deployment in the CI lab does not support access to
glance.ovirt.org. To be able to run the suite, a local image is uploaded
to the engine with imageio instead.
This upload will become the default for both IPv4/IPv6.
'image_template' role by default creates a first vnic connected to
ovirtmgmt inside the template. This is not expected in network-suite
tests which configure VMs based on a vanilla template. Therefore,
specify no nics in the upload.

Change-Id: I33bb670e1e69ae51eea1bf778ca6f79bf50cb000
Signed-off-by: Eitan Raviv <eraviv@redhat.com>