he: test_012_local_maintenance_sdk: Fix de/activation loops
It seems like activating an already-active host puts it momentarily to
unassigned state. Recently we had more failures due to this. Not sure if
this is a result of changes on the engine side. Anyway, split the single
loop we had per each step with two: One to de/activate, other to check
if it's in/active.

Change-Id: I5b4645b81793d35a3fd3034633934976e9e6d030
Signed-off-by: Yedidyah Bar David <didi@redhat.com>