make sure update template version is done before update vm pool
We need to make sure that the VMs in the VM pool are updated to
the latest template version before updating the VM pool, otherwise
the template might be locked. This patch also changes the VM that
we create the template version from to the vm imported from OVF
as it contains a single disk (rather than three).

Change-Id: Ib9b304b06ec101b97c0972bd2be70f89598386b4
Signed-off-by: Arik Hadas <ahadas@redhat.com>