support github url with multiple artifacts
Currently, OST custom repos can't be used with GitHub URLs
that created multiple distro artifacts.
Since we use the first rpm in the list created by the github actions,
if the ost image distro does not match this rpm distro,OST will fail.
we should condition to match the right artifacts to the
image distribution.

Change-Id: Ib9579896b3ac8aa39be50efaa0abe4bb0a5376b2
Signed-off-by: Harel Braha <hbraha@redhat.com>