basic: Remove time sync
Syncing hosts' time with the engine does not really work, since
using the command tool is not persistent and the hosts can't connect
to engine without permission.

Since all the VMs use other servers for syncing anyway, it's unlikely
that this will result in bigger time differences anyway, so we can
just drop the code instead of fixing it.

Change-Id: I5c9987aef6bd532842fb7e7b7335e425bcfc9fa3
Signed-off-by: Andrej Cernek <acernek@redhat.com>