ost_utils: backend - rename ost_name to network_role
Networks created at layer 0 by the ost scripts have a 'network-type'
attribute assigned in the meta-data of their describing XML.
The possible values for this attribute are 'management', 'storage' and
'bonding' and the networks are used according to these names. So a
better descriptor for them in the backend is 'network_role'.

Change-Id: I1d387da1a8e4007b93e4b4d67bddaf1ec7bc52cc
Signed-off-by: Eitan Raviv <eraviv@redhat.com>