ost_utils: nic mac address to network role mapping
The deployment network XMLs specifies:
- the MAC address for each NIC in the host VMs
- the MAC address of the corresponding veth device and the network which
  is plugged into that veth

This data enables ost_utils/backend to correlate a host NIC to the
underlying network using the network role naming convention.

This correlation is now exposed as a mapping per host in order for test
code to be able to map a NIC's name to the network role. As a result,
host NIC names no longer need to be hard-coded in test code.

Change-Id: If8fe994468159611d4b20af4c33fab53b088fd55
Signed-off-by: Eitan Raviv <eraviv@redhat.com>