basic: 098 - wait for provider sync completion event
Engine periodically invokes a sync request to fetch updates from ovn
provider. In test_use_ovn_provider the sync is disabled as part of the
test but if a sync request has not completed before the disable then a
race situation is created which causes failures at various stages in the
test.
To avoid such a race, a wait on the completion of the sync action is
added after disabling it. Either the sync has completed before the wait
or the wait should catch it as 10 seconds is long enough for the whole
round trip of the sync request.

Change-Id: I68dfd4512ccd5de2c1e72b0f638a10b101b86652
Signed-off-by: Eitan Raviv <eraviv@redhat.com>