lagofy: add OST_DEPLOYMENT environment variable
New environment variable OST_DEPLOYMENT can be used to configure a
different location for deployments. Normally this is in deployment/
subdirectory of the OST repository. But in certain scenarios, e.g. for
storage reasons or to avoid permission issues when repository is on NFS,
it makes sense to provide different location outside of the repository
tree.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>