Update Grafana check for error
After the change in the Grafana UI we are no longer
able to detect error messages displayed.

This patch updates the checks for errors in Grafana and
also makes sure that the test fails if the notifications
list element is not detected - that could mean that the UI changed
and we are not able to detect errors again.