Check VM Portal status based on presence of Refresh button
VM Portal uses an internal flag appReady to check if the app has been
fully initialized. The flag is used to toggle features that require
fully operational app. One of such features is manual refresh.

The test can indirectly check for appReady flag by checking if Refresh
button is displayed in the toolbar.

Reference-Url: https://github.com/oVirt/ovirt-web-ui/blob/bfd825b22b9b4acc14c3309f2832c1f8bd9e6e28/src/components/VmsPageHeader/index.js#L65
Change-Id: If84da15b65c2ede2817770eeb75a8452262e4f44
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>