metrics: Use daemonic threads to run the playbooks
The PR is a continuation of the effort to mitigate problems with metrics
configuration playbook hanging forever. The first part was here [1].

Now we don't hang on the playbook execution, but we do hang at the end
of the pytest session - Python waits on all the threads to be done.
That rule doesn't apply to daemonic threads, so let's use them in this
case.

[1] https://github.com/oVirt/ovirt-system-tests/pull/188

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>