he: skip sar installation on el9stream hosts
During ost-images build we cache sar RPMs so they can be installed
offline on the HE. We don't have an el9stream based HE yet, so the
RPMs cached on el9stream hosts are not installable in this case.
Let's skip the sar installation for now when running on el9stream.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>