engine: Restore original 'ovirt-engine.xml.in' after engine setup
PCI DSS openscap profile verifies checksums of files coming from RPMs.
Modified files are reported as a problem. We change the contents
of 'ovirt-engine.xml.in' file shipped as a part of 'ovirt-engine-backend'
RPM to enable debug logging on the engine.

Let's backup the file before changing it and restore the original after
engine setup.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>