Improve disk_template driver configuration
Use the same configuration for the disk_template as we use for the root
disk in vm_template. This is more important since the storage vm does
more I/O.

The disks do not use iothread='1' since they use bus='scsi', and in this
case iothread is set on the controller. However we don't have a scsi
controller in the vm template yet.

Fixes e2453ad536ef (vm-template: Improve disk driver configuration)