UI tests: Move directly to welcome screen after logout
After the logout, the behavior is not deterministic. Sometimes
is the user redirected to welcome screen, sometimes to login screen
and somtimes it stays on the Webadmin dashboard for a while.

This patch navigates to the welcome screen after pressing log out,
and if necessary, reloads and waits until the user is logged out.