basic: Check TPM presence
We enable TPM for vm0 and then check that a TPM device is actually
present in the guest OS.  TPM with FIPS is broken in CentOS Stream 8,
we enable the tests only on the other distros (RHEL, CentOS Stream 9).

We must use backup_vm for the test rather than vm0 because vm0 tests a
snapshot with memory, which is incompatible with TPM.  backup_vm
configuration must be changed to a TPM-compatible guest OS in order to
be able to have a TPM device.