Fix selenium timeouts
From time to time Selenium server hangs indefinitely on http requests
made by client. Looking at the logs it seems that the netty code sees
the http request, but somehow it never reaches the code that handles the
response.

The client side blocks until it gets a response from the server.
The timeout for this path seems to be 5 mins. On the server side
however, the timeout for an inactive session to be killed is set
to 3 mins by default. That combination creates an undesired effect -
before the client realizes that the http request may have been stuck for
too long, the server already kills its session.

This patch shortens the client timeout for waiting on http
responses to 30 secs. Unfortunately, with current client implementation
in Selenium library, it's not possible to change this value
through i.e. an '__init__' argument - it rather enforces subclassing.
Although ugly, changing the 'RemoteConnection._timeout' class member
directly works for us.

Furthermore, the recent client libraries does not seem to translate an
urllib3's 'ReadTimeoutError' exception to Selenium's 'TimeoutException'
This is why we also add it to a list of handled/ignored exceptions
in a couple of places.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>