Dump console and firmware logs for all VMs
Problems with nested VMs' firmware are hard to debug. A dump of firmware
log is very helpful with that, so let's have it on by default in OST.

Initially the console logs were only set up for HE, but having them
for all the VMs helps a lot too.

Needs https://github.com/oVirt/ost-images/pull/91 to work properly.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>