ci: run ost on self-hosted github runner
Run ost on a self-hosted runner because the old Jenkins infrastructure
is no longer available. Usage of the workflows, the way you trigger ost
etc all stayed the same as before (commenting `/ost ...`)

NOTE: a breaking change is that the "trigger" workflow now needs to define
permissions, see `.github/workflows/example-trigger.yml`. This change
will need to be applied to repos that need to call ost.

- Used `pytest-md-report` to add report to workflow summary.
- Deployment dir is put outside of the workflow working dir, because the
  workflow will be reused from multiple repositories.
- A previous deployment is destroyed at the start of the workflow. This
  allows for debugging the infra should it be necessary.
- "exported-artifacts" are uploaded to GitHub and downloadable on the
  Summary page of the workflow.
- The workflow comments on PRs about start/success/failure that triggered
  ost.

Signed-off-by: Peter Boden <peter.boden@team.blue>